requirements.txt
flask
firebase-admin
deepface
pandas
pillow
